Add files via upload
This repository consists of the Master and Transaction tables for the user management project of a vedio streaming platform(basic). The technologies used here are HTML, CSS for the Front end and Ux part. Oracle Liv Sql for the Database part. JavaScript for the integration. This is my first project in full stack and Im working to improve it to make it more deployement friendly and more interactive.